Create the Code
This program first imports urllib, json, and ssl. As a default, ‘api_key’ is set to False. If the user does not have an api key, then ‘api_key’ is False and the program uses the course-given api_key and service url. If the user has an api key for Google Maps, api_key should be set to True and the code can use the included google maps service url.

The code prompts for a location/address from the user and if the user does not add a url (<1 character in length), uses the included address/location. Next it makes a dictionary that contains the address, so urlencode accepts it and then it concatenates the service url and code for the encoded location address. 
    
The code then prints the newly concatenated url, opens the url and reads and decodes the data. It tries to parse the data and extracts the ‘place_id’ data.